Finished login, added interceptors and alert service

+++ frontend/hr-system/src/app/components/projects/all-projects/all-projects.component.ts
@@ -3,6 +3,8 @@ import { Project } from '../../../core/models/project.model';
 import { ProjectService } from '../../../core/services/project.service';
 import { Router } from '@angular/router';
 import { Subject, switchMap, takeUntil, tap } from 'rxjs';
+import Swal from 'sweetalert2'
+import { AlertService } from '../../../core/services/alert.service';
 
 @Component({
   selector: 'app-all-projects',
@@ -18,7 +20,7 @@ export class AllProjectsComponent implements OnInit, OnDestroy {
 
   private destroy$ = new Subject<void>();
 
-  constructor(private projectService: ProjectService, private router: Router) {}
+  constructor(private projectService: ProjectService, private router: Router, private swal :AlertService) {}
 
   ngOnInit() {
     this.projectService.getAllProjects(this.getQueryString())
@@ -48,13 +50,26 @@ export class AllProjectsComponent implements OnInit, OnDestroy {
   }
 
   delete(id : string) : void {
-    console.log("deleted")
-    this.projectService.deleteProject(id).pipe(takeUntil(this.destroy$),
-     switchMap( () => this.projectService.getAllProjects(this.getQueryString()).pipe(
-       tap(response => {
-         this.projects = response.projects;
-         this.totalPages = response.totalPages;
-       })))).subscribe()
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.projectService.deleteProject(id).pipe(takeUntil(this.destroy$),
+          switchMap( () => this.projectService.getAllProjects(this.getQueryString()).pipe(
+          tap(response => {
+            this.projects = response.projects;
+            this.totalPages = response.totalPages;
+          })))).subscribe()
+        this.swal.fireSwalSuccess("Project deleted successfully")
+      }
+    });
+    
   }
 
   ngOnDestroy(): void {
